[C] add marker formatting
