+++ package.json
@@ -47,7 +47,7 @@
   },
   "dependencies": {
     "@castiron/style-mixins": "^1.0.6",
-    "@rubin-epo/epo-react-lib": "^1.2.5",
+    "@rubin-epo/epo-react-lib": "^1.2.6",
     "@rubin-epo/epo-widget-lib": "^0.2.1",
     "classnames": "^2.3.2",
     "focus-trap": "^7.3.1",
